+++ contracts/Registry.sol
@@ -3,11 +3,11 @@ pragma solidity ^0.4.24;
 
 contract PSPRegistry {
 
-    event PSPAdded(string psp_name, address psp_address);
+    event PSPAdded(bytes32 psp_name, address psp_address);
 
     address public owner;
     mapping (bytes32 => address) nameToAddress;    // mapping from psp name to psp ERC 725 contract address
-    string[] public registeredPSP;
+    bytes32[] public registeredPSP;
 
     modifier onlyOwner {
         require(msg.sender == owner);
@@ -18,11 +18,11 @@ contract PSPRegistry {
         owner = msg.sender;
     }
 
-    function getPSPAddress(string psp_name) public view returns (address psp_address) {
+    function getPSPAddress(bytes32 psp_name) public view returns (address psp_address) {
         return nameToAddress[keccak256(abi.encodePacked(psp_name))];
     }
 
-    function registerPSP(string psp_name, address psp_address) onlyOwner public {
+    function registerPSP(bytes32 psp_name, address psp_address) onlyOwner public {
         nameToAddress[keccak256(abi.encodePacked(psp_name))] = psp_address;
         registeredPSP.push(psp_name);