+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # June 2018
 
+
+- v0.1 psp key
+  - Created registry contract
+  - Created PSP_Identity contract
+  - Added executePayment to Identity ☢️ double check right needed
+  - Updated tests
+
 - Updated to solc 0.4.24
   - constructor changed
   - emit key word added when dispatching an event